DINARCOIN DC
DinarCoin is DinarCoi Dc ’s Crypto Asset, created on Bainans Smart Blockchain, being pegged to a gold spot contract (BNB) from the BNB liquidity provider.

The total amount of DinarCoin (DC) in circulation can be verified on Blockchain or checking the BNB held in escrow at BNB liquidity provider. The backbone of the technology behind is the Public Bainans’s Blockchain mechanism.

Every DC is pegged to real, physical gold, so the risk of inflation is eliminated. We are working hard to make DC to be the perfect currency: spendable, anonymous, theft resistant, open source, non-counterfeitable, durable, inflation proof, decentralized and scarce. Users can trade DNC, convert it into real money (dollars, euro, & etc.), real physical gold or use it as payments at certain merchants that DinarCoin Dc will engage soon in 2022. This is the beginning of a new era of global currencies. However DC is created as the 1st Crypto Asset backed by gold in the world.

DC could easily be transferred or sell this token and receive cash settlement within less than a minutes using the Public Bainans Blockchain Technology. Hence this will eventually removing storage costs, direct taxation and generally all the headaches of the commodity-to-cash conversion process.
DC are the form of Digital Currency (again each one backed by gold). They can be traded, used to purchase items, and used for just about any other use a digital currency can be used for instance that already in the market, Bitcoin btc, Bnb,Busd, Cake . Banana & Bitcoin btcmt. As a conclusion, DC may be purchased or “minted” (which will actually create new DC backed by physical gold).

STRUCTURE OF BNB PRODUCT

DinarCoin can be procured from a holder of DinarCoin on the open market, or they can procure freshly minted new DC from liquidity provider. On the other hand, DinarCoin BNB (Decentralized Autonomous Organization) can mint new DC by procuring BNB to the 1:1 value of the DC through BNB Master Account. For the lifetime of a DC a corresponding value of BNB will be held in escrow that will STP 1:1 to the liquidity provider and available to be liquidized anytime when close the orders. Furthermore, when DC is created it is registered on the Bainans Blockchain and the Bitcoin Blockchain. The total amount of DC in circulation can be verified on either Blockchain, and audited against the total BNB held in escrow/Liquidity Provider by DinarCoin.

Website : https://buy.dinarcoindc.cc

Official website : https://buy.dinarcoindc.cc